wgengine/magicsock: extract IMDS utilities into a standalone package
Moves magicksock.cloudInfo into util/cloudinfo with minimal changes.

Updates #17796
